+++ test/helpers/FileList.js
@@ -3,26 +3,36 @@
 //   https://github.com/jsdom/jsdom/issues/1272
 
 /**
- * @class
+ * `FileList` polyfill.
  */
-function _FileList () {
-    // eslint-disable-next-line prefer-rest-params -- API
-    this._files = arguments[0];
-    this.length = this._files.length;
+// eslint-disable-next-line sonarjs/class-name -- Avoid shadowing global
+class _FileList {
+    /**
+     *
+     */
+    constructor () {
+        // eslint-disable-next-line prefer-rest-params -- API
+        this._files = arguments[0];
+        this.length = this._files.length;
+    }
+
+    /**
+     * @param {import('typeson').Integer} index
+     * @returns {File}
+     */
+    item (index) {
+        return this._files[index];
+    }
+
+    /* eslint-disable class-methods-use-this -- Not needed */
+    /**
+     * @returns {"FileList"}
+     */
+    get [Symbol.toStringTag] () {
+        /* eslint-enable class-methods-use-this -- Not needed */
+        return 'FileList';
+    }
 }
-
-/**
- * @param {import('typeson').Integer} index
- * @returns {File}
- */
-_FileList.prototype.item = function (index) {
-    return this._files[index];
-};
-
-// eslint-disable-next-line unicorn/no-top-level-side-effects -- Ok
-Object.defineProperty(_FileList.prototype, Symbol.toStringTag, {
-    value: 'FileList'
-});
 
 // eslint-disable-next-line unicorn/no-top-level-side-effects -- Ok
 Object.defineProperty(globalThis.HTMLInputElement.prototype, 'files', {